lab01/requirements.txt
faker>=19.0.0
psycopg2-binary>=2.9.0
matplotlib>=3.5.0
seaborn>=0.11.0
pandas>=1.5.0
numpy>=1.21.0
